Fix sonar warnings
