+++ src/datascience-ui/native-editor/nativeEditor.tsx
@@ -8,6 +8,7 @@ import { OSType } from '../../client/common/utils/platform';
 import { concatMultilineStringInput } from '../../client/datascience/common';
 import { NativeCommandType } from '../../client/datascience/interactive-common/interactiveWindowTypes';
 import { ContentPanel, IContentPanelProps } from '../interactive-common/contentPanel';
+import { handleLinkClick } from '../interactive-common/handlers';
 import { ICellViewModel, IMainState } from '../interactive-common/mainState';
 import { IStore } from '../interactive-common/redux/store';
 import { IVariablePanelProps, VariablePanel } from '../interactive-common/variablePanel';
@@ -340,20 +341,7 @@ export class NativeEditor extends React.Component<INativeEditorProps> {
     }
 
     private linkClick = (ev: MouseEvent) => {
-        // If this is an anchor element, forward the click as Jupyter does.
-        let anchor = ev.target as HTMLAnchorElement;
-        if (anchor && anchor.href) {
-            // Href may be redirected to an inner anchor
-            if (anchor.href.startsWith('vscode')) {
-                const inner = anchor.getElementsByTagName('a');
-                if (inner && inner.length > 0) {
-                    anchor = inner[0];
-                }
-            }
-            if (anchor && anchor.href && !anchor.href.startsWith('vscode')) {
-                this.props.linkClick(anchor.href);
-            }
-        }
+        handleLinkClick(ev, this.props.linkClick);
     }
 
 }
